fix(governance): decouple buildVote unit test from staking-sdk ESM chain
The vote-mapping unit test imported buildVote from vote.ts, which pulls in
the Pyth staking client and its transitive ESM-only dependency
(@pythnetwork/client -> camelcase@8), which Jest's bare defineJestConfig()
cannot require(). Isolate buildVote and the VoteSide type into vote-side.ts,
which imports only @solana/spl-governance, so the test no longer drags in the
staking client and runs on a clean build.

Also harden the execute path: guard against the signed blockhash expiring
while waiting for a slow Fireblocks approval (report an actionable error
instead of an opaque send failure), reject negative ledger derivation values,
and document the prompt-approval requirement in the README.
